fmna: intialization: check if BLE stack is enabled
Added a check to verify if the BLE stack is enabled during the FMN
initialization. It is necessary to have the BLE stack enabled because
the FMN stack uses it. It is not possible to enable BLE inside the FMN
because the BLE API is used to obtain values for FMN initialization
parameters.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>