doc: add documentation build system and integrate existing content
Added a documentation build system for the Find My repository. This
system is built on top of the nRF Connect SDK documentation framework
that has been modified to accomodate the Find My requirements.
Necessary modifications are kept in the nRF Connect SDK fork:

https://github.com/kapi-no/fw-nrfconnect-nrf/tree/find_my_doc_fork

Generated documentation in the HTML format is kept on the special
documentation branch in the Find My repository:

https://github.com/nrfconnect/sdk-find-my/tree/doc_artifacts

This change also integrates existing documentation content for Find My
with the new build system.

Added RST pages with release notes.

Ref: NCSDK-9576

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>