fmna: track maintained accessory state & encode it in the adv payload
Added logic to the FMN State module for tracking the maintained status
of the accessory. This status flag is part of advertising payload for
Nearby and Separated advertising.

This change also adds configuration for this flag to the FMN
Advertising module internal API and encoding logic in the advertising
code.

Ref: NCSDK-9178

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>