fmna: conn: add authentication restore logic
Added authentication callback management logic to the FMN Connection
module. It resets the authentication callbacks in the Bluetooth LE
stack for Find My peers so that they can use Just Works pairing. Then,
the authentication callbacks are restored in the Bluetooth LE stack so
that it is possible for non-FMN identities to be paired using any
pairing method (e.g. Passkey display).

Ref: NCSDK-9920

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>