fmna: add location availability callback for device naming
Added a callback to indicate whether or not the location of the
accessory is available to non-owner devices from the Find My Network.
This API is intended only for "pair before use" accessories.
It is used to determine if the "Find My" suffix should be appended to
the device name for their primary purpose Bluetooth activity
(e.g. advertising, device name GATT characteristic).

Ref: NCSDK-10403

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>