fmna: serial_number: fix write outside array boundary
The fmna_serial_number_get API was unsafe to use. In certain cases
the implementation could write data to the input parameter outside
its memory boundary.

This array overwrite could happen in the bin2hex function because
this function puts string terminator character after encoding is done.
If all 16 bytes of the array are used for the Serial Number storage,
then the bin2hex will write one additional byte right after the array
memory region.

One of the consequences of this bug was the incorrect encoding of the
encrypted Serial Number. The counter parameter, located right after
the serial number character string, was reset to zero by this array
overwrite.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>