samples: common: increase BT thread stack for release and debug variant
Increased the stack size of the Bluetooth RX thread to avoid stack
overflow during Find My pairing. The configuration is changed in all
Find My samples and applications for both the release and the debug
variant. Stack of the Bluetooth RX thread is used in the characteristic
write callback of the Find My service control points.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>