fmna: serial_number: fix bin2hex conversion during SN generation
Fixed the conversion from the binary data to character string during
serial number generation in Find My. The bin2hex implementation is
fixed in the Zephyr upstream and does not require a workaround to avoid
memory overwrite anymore. Now, it is okay to pass the actual size of the
output array.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>