fmna: serial_number: add support for non-secure target on nRF5340
Aligned the Find My Serial Number module to support non-secure target.
The module uses Device ID from the FICR register to generate serial
number in the default mode.

To fetch Device ID from the secure context, it is necessary either to
use the Secure Services API in the legacy SPM non-secure build or to
use TF-M API in the TF-M enabled non-secure build.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>