fmna: add disable and ready API to the stack
Added the fmna_disable function to the public Find My API.
This function deactivates the FMN feature. As a result, all Find My
functionalities like advertising, NFC emulation and key rotation are
stopped. During the disabling process, the accessory also disconnects
all Find My peers that are connected to it over Bluetooth. The disabled
state of the FMN stack is treated similarly to the power-off state.

Added the fmna_is_ready function to the public Find My API. This
function can be used to check if the FMN stack is enabled or disabled.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>