fmna: state: fix for unpair command rejection
Fixed an issue with the Unpair command rejection when the command
appeared in the same connection session as the Find My pairing.

The issue caused the FMN stack to persist in the paired state when
the Find My app user cancelled the pairing flow after a successful
pairing packet exchange.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>