tools: ncsfmntools: Fix problem with data extraction from settings
In case of multiple settings partition sectors and multiple ex. tokens
entries, there was a chance that wrong data has been extracted.
This change modifies script in a way that is aware which one sector is
currently used and iterates through memory until it finds the right one.

Jira: NCSDK-20053

Signed-off-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>