fmna: motion_detection: Fix timer period comparison
Fixed an invalid timer period comparison in the Motion Detection
module that caused Unwanted Tracking Detection to work incorrectly.

Before fix in Zephyr kernel, timer period was modified by subtracting
1 in k_timer_start() function. Commit
https://github.com/zephyrproject-rtos/zephyr/commit/6a51a10decf92a2a0d585d485aada065877625c8
changed this behavior and now period that is passed to timer expiry_fn
is equal to the one passed as argument to start function and there
is no need to manually subtract 1 anymore when comparing.

Jira: NCSDK-22173

Signed-off-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>