fmna: add API for managing Find My paired advertising
Removed the automatic mechanism in the Find My library that pauses the
Find My paired advertising once a Bluetooth peer connects to the
device using other Bluetooth identity than the Find My identity.
The Find My paired advertising was also automatically resumed once
all such peers disconnect from the device.

Added two new Find My APIs that manage the paired advertising. The APIs
are symmetrical: one is used to enable the paired advertising and the
other is used to disable it. By default, the paired advertising is
enabled.

Aligned the Find My Coexistence sample to demonstrate the new API,
which replaces the automatic mechanism from the Find My library.

The described feature is necessary for pair before use accessories.
The automatic mechanism is replaced by the new APIs to provide more
flexibility for the developers.

Ref: NCSDK-18006

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>